extract should-build? into separate script
